--- src/App.js
@@ -1,11 +1,0 @@
-import './styles/App.css';
-
-function App() {
-  return (
-    <>
-      <h1>React App</h1>
-    </>
-  );
-}
-
-export default App;
