+++ src/components/Details.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Details = () => (
+  <div>Details</div>
+);
+
+export default Details;
