+++ src/components/Home.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Home = () => (
+  <div>Home</div>
+);
+
+export default Home;
